+++ addons/Swift_Inventory/Classes/SwiftGrid.gd
@@ -7,7 +7,6 @@
 class_name SwiftGrid
 extends SwiftContainer
 
-
 ## Number of addresses displayed by the grid.
 ##
 ## This property proxies [member SwiftInventory.size] on [member swift_inventory].
@@ -39,62 +38,43 @@ func _validate_property(property: Dictionary) -> void:
 		property.usage &= ~PROPERTY_USAGE_EDITOR
 
 
-func _on_swift_change_add(address: int) -> void:
-	_refresh_slot(address)
-
-func _on_swift_change_remove(address: int) -> void:
-	_refresh_slot(address)
-
-func _on_swift_change_move(from_address: int, to_address: int) -> void:
-	_refresh_slot(from_address)
-	_refresh_slot(to_address)
-
-func _on_swift_change_swap(first_address: int, second_address: int) -> void:
-	_refresh_slot(first_address)
-	_refresh_slot(second_address)
-
-func _on_swift_change_transfer(from_address: int, to_address: int) -> void:
-	_refresh_slot(from_address)
-	_refresh_slot(to_address)
-
-func _on_swift_change_set(address: int) -> void:
-	_refresh_slot(address)
-
-func _on_swift_change_size() -> void:
-	if not is_node_ready():
-		return
-	_sync_slots()
-	_refresh_slots()
-
-func _on_swift_change_inventory() -> void:
-	_refresh_slots()
-
-
-func _sync_slots() -> void:
-	while get_child_count() > inventory_size:
-		_remove_slot(-1)
-	while get_child_count() < inventory_size:
-		_add_slot(get_child_count())
-	for index in get_child_count():
-		var slot := get_child(index) as SwiftSlot
-		if slot:
+func _reconcile_all() -> void:
+	var target_size := inventory_size
+	var slots := _get_slot_children()
+	while slots.size() > target_size:
+		var slot := slots.pop_back()
+		_remove_slot_node(slot)
+	while slots.size() < target_size:
+		slots.append(_create_slot(slots.size()))
+	for address in slots.size():
+		var slot := slots[address]
+		if slot.texture_rect == null or slot.amount_label == null:
 			slot.setup()
-			slot.bind(swift_inventory, index)
+		slot.name = ("Slot_%s" % address).pad_zeros(2)
+		slot.size = slot_size
+		slot.bind(swift_inventory, address)
 	queue_sort()
 
 
+func _reconcile_address(address: int) -> void:
+	if address < 0 or address >= inventory_size:
+		return
+	var slots := _get_slot_children()
+	if (
+		slots.size() != inventory_size
+		or slots[address].address != address
+		or slots[address].swift_inventory != swift_inventory
+	):
+		_reconcile_all()
+		return
+	slots[address].refresh()
+
+
 func _sort_slots() -> void:
-	var columns := maxi(1, floori((size.x + separation.x) / (slot_size.x + separation.x)))
+	var stride := slot_size + separation
+	var columns := maxi(1, floori((size.x + separation.x) / stride.x))
+	var slots := _get_slot_children()
 
-	for i in get_child_count():
-		var slot := get_child(i) as SwiftSlot
-		if not slot:
-			continue
-
-		var column := i % columns
-		var row := i / columns
-		var slot_position := Vector2(
-			column * (slot_size.x + separation.x), row * (slot_size.y + separation.y)
-		)
-
-		fit_child_in_rect(slot, Rect2(slot_position, slot_size))
+	for index in slots.size():
+		var slot_position := Vector2(index % columns, index / columns) * Vector2(stride)
+		fit_child_in_rect(slots[index], Rect2(slot_position, slot_size))